+++ .eslintrc.yml
@@ -1,4 +1,9 @@
+parserOptions:
+  ecmaVersion: 2015
+  sourceType: 'module'
+
 env:
+  node: true
   es6: true
   browser: true
   commonjs: true
